requirements.txt
zeroconf
paho-mqtt
